requirements.txt
PyGithub
bs4
python-dotenv
discord-py-slash-command
discord.py
selenium